refact command
