+++ src/main.rs
@@ -16,19 +16,16 @@ use clap::{App, Arg, ArgMatches};
 use colored::{self, *};
 use dirs::home_dir;
 use log::LevelFilter;
-use regex::Regex;
 
-use command::{print_usage, secret};
+use command::{common_usage, do_command, secret};
 use config::Config;
-use logger::{change_log_type, Logger};
+use logger::Logger;
 
 mod command;
 mod config;
 mod logger;
 
 static MY_LOGGER: Logger = Logger;
-static S3_FORMAT: &'static str =
-    r#"[sS]3://(?P<bucket>[A-Za-z0-9\-\._]+)(?P<object>[A-Za-z0-9\-\._/]*)"#;
 static S3RS_CONFIG_FOLDER: &'static str = ".config/s3rs";
 
 fn read_parse<T>(tty: &mut File, prompt: &str, min: T, max: T) -> io::Result<T>
@@ -69,225 +66,6 @@ fn my_pick_from_list_internal<T: AsRef<str>>(items: &[T], prompt: &str) -> io::R
     Ok(idx)
 }
 
-fn print_if_error(result: Result<(), failure::Error>) {
-    match result {
-        Err(e) => println!("{}", e),
-        Ok(_) => {}
-    };
-}
-
-fn do_command(handler: &mut s3handler::Handler, s3_type: &String, command: &mut String) {
-    debug!("===== do command: {} =====", command);
-    if command.starts_with("la") {
-        match handler.la() {
-            Err(e) => println!("{}", e),
-            Ok(v) => {
-                for o in v {
-                    debug!("{:?}", o);
-                    println!("{}", String::from(o));
-                }
-            }
-        };
-    } else if command.starts_with("ls") {
-        match handler.ls(command.split_whitespace().nth(1)) {
-            Err(e) => println!("{}", e),
-            Ok(v) => {
-                for o in v {
-                    debug!("{:?}", o);
-                    println!("{}", String::from(o));
-                }
-            }
-        };
-    } else if command.starts_with("ll") {
-        let r = match command.split_whitespace().nth(1) {
-            Some(b) => handler.ls(Some(b)),
-            None => handler.la(),
-        };
-        match r {
-            Err(e) => println!("{}", e),
-            Ok(v) => {
-                println!("STORAGE CLASS\tMODIFIED TIME\t\t\tETAG\t\t\t\t\tKEY",);
-                for o in v {
-                    debug!("{:?}", o);
-                    println!(
-                        "{}\t{}\t{}\t{}",
-                        o.storage_class.clone().unwrap_or("        ".to_string()),
-                        o.mtime
-                            .clone()
-                            .unwrap_or("                        ".to_string()),
-                        o.etag
-                            .clone()
-                            .unwrap_or("                                 ".to_string()),
-                        String::from(o)
-                    );
-                }
-            }
-        };
-    } else if command.starts_with("put") {
-        match handler.put(
-            command.split_whitespace().nth(1).unwrap_or(""),
-            command.split_whitespace().nth(2).unwrap_or(""),
-        ) {
-            Err(e) => println!("{}", e),
-            Ok(_) => println!("upload completed"),
-        };
-    } else if command.starts_with("get") {
-        match handler.get(
-            command.split_whitespace().nth(1).unwrap_or(""),
-            command.split_whitespace().nth(2),
-        ) {
-            Err(e) => println!("{}", e),
-            Ok(_) => println!("download completed"),
-        };
-    } else if command.starts_with("cat") {
-        if let Ok(o) = handler.cat(command.split_whitespace().nth(1).unwrap_or("")) {
-            println!("{}", o.1.unwrap_or("".to_string()));
-        } else {
-            error!("can not cat the object");
-        }
-    } else if command.starts_with("del") || command.starts_with("rm") {
-        let mut iter = command.split_whitespace();
-        let target = iter.nth(1).unwrap_or("");
-        let mut headers = Vec::new();
-        loop {
-            match iter.next() {
-                Some(header_pair) => match header_pair.find(':') {
-                    Some(_) => headers.push((
-                        header_pair.split(':').nth(0).unwrap(),
-                        header_pair.split(':').nth(1).unwrap(),
-                    )),
-                    None => headers.push((&header_pair, "")),
-                },
-                None => {
-                    break;
-                }
-            };
-        }
-        match handler.del_with_flag(target, &mut headers) {
-            Err(e) => println!("{}", e),
-            Ok(_) => println!("deletion completed"),
-        }
-    } else if command.starts_with("tag") {
-        let mut iter = command.split_whitespace();
-        let action = iter.nth(1).unwrap_or("");
-        let target = iter.nth(0).unwrap_or("");
-        let mut tags = Vec::new();
-        loop {
-            match iter.next() {
-                Some(kv_pair) => match kv_pair.find('=') {
-                    Some(_) => tags.push((
-                        kv_pair.split('=').nth(0).unwrap(),
-                        kv_pair.split('=').nth(1).unwrap(),
-                    )),
-                    None => tags.push((&kv_pair, "")),
-                },
-                None => {
-                    break;
-                }
-            };
-        }
-        match action {
-            "add" | "put" => match handler.add_tag(target, &tags) {
-                Err(e) => println!("{}", e),
-                Ok(_) => println!("tag completed"),
-            },
-            "del" | "rm" => match handler.del_tag(target) {
-                Err(e) => println!("{}", e),
-                Ok(_) => println!("tag removed"),
-            },
-            "ls" | "list" => match handler.list_tag(target) {
-                Err(e) => println!("{}", e),
-                Ok(_) => {}
-            },
-            _ => println!("only support these tag actions: ls, add, put, del, rm"),
-        }
-    } else if command.starts_with("usage") {
-        let mut iter = command.split_whitespace();
-        let target = iter.nth(1).unwrap_or("");
-        let mut options = Vec::new();
-        loop {
-            match iter.next() {
-                Some(kv_pair) => match kv_pair.find('=') {
-                    Some(_) => options.push((
-                        kv_pair.split('=').nth(0).unwrap(),
-                        kv_pair.split('=').nth(1).unwrap(),
-                    )),
-                    None => options.push((&kv_pair, "")),
-                },
-                None => {
-                    break;
-                }
-            };
-        }
-        match handler.usage(target, &options) {
-            Err(e) => println!("{}", e),
-            Ok(_) => {}
-        }
-    } else if command.starts_with("mb") {
-        print_if_error(handler.mb(command.split_whitespace().nth(1).unwrap_or("")));
-    } else if command.starts_with("rb") {
-        print_if_error(handler.rb(command.split_whitespace().nth(1).unwrap_or("")));
-    } else if command.starts_with("/") {
-        match handler.url_command(&command) {
-            Err(e) => println!("{}", e),
-            Ok(_) => {}
-        };
-    } else if command.starts_with("info") {
-        let target = command.split_whitespace().nth(1).unwrap_or("");
-        let caps;
-        let bucket = if target.starts_with("s3://") || target.starts_with("S3://") {
-            let re = Regex::new(S3_FORMAT).unwrap();
-            caps = re
-                .captures(command.split_whitespace().nth(1).unwrap_or(""))
-                .expect("S3 object format error.");
-            &caps["bucket"]
-        } else {
-            target
-        };
-        println!("{}", "location:".yellow().bold());
-        let _ = handler.url_command(format!("/{}?location", bucket).as_str());
-        println!("\n{}", "acl:".yellow().bold());
-        let _ = handler.url_command(format!("/{}?acl", bucket).as_str());
-        println!("\n{}", "versioning:".yellow().bold());
-        let _ = handler.url_command(format!("/{}?versioning", bucket).as_str());
-        match s3_type.as_str() {
-            "ceph" => {
-                println!("\n{}", "version:".yellow().bold());
-                let _ = handler.url_command(format!("/{}?version", bucket).as_str());
-                println!("\n{}", "uploads:".yellow().bold());
-                let _ = handler.url_command(format!("/{}?uploads", bucket).as_str());
-            }
-            "aws" | _ => {}
-        }
-    } else if command.starts_with("s3_type") {
-        handler.change_s3_type(&command);
-    } else if command.starts_with("auth_type") {
-        handler.change_auth_type(&command);
-    } else if command.starts_with("format") {
-        handler.change_format_type(&command);
-    } else if command.starts_with("url_style") {
-        handler.change_url_style(&command);
-    } else if command.starts_with("log") {
-        change_log_type(&command);
-    } else if command.starts_with("exit") || command.starts_with("quit") {
-        println!("Thanks for using, cya~");
-    } else if command.starts_with("help") {
-        println!(
-            r#"
-USAGE:"#
-        );
-        print_usage();
-        secret::print_usage();
-        println!(
-            "If you have any issue, please submit to here https://github.com/yanganto/s3rs/issues"
-        );
-    } else {
-        println!(
-            "command {} not found, help for usage or exit for quit",
-            command
-        );
-    }
-}
 fn cli() -> ArgMatches<'static> {
     App::new(env!("CARGO_PKG_NAME"))
         .version(env!("CARGO_PKG_VERSION"))
@@ -308,7 +86,7 @@ fn cli() -> ArgMatches<'static> {
                 .help("Set the run time secret to encrypt/decrept your s3config file")
 		)
         .arg(
-            Arg::from_usage("<s3rs cmd>...").requires("config").required(false)
+            Arg::from_usage("<s3rs cmd>...").help(&common_usage()).requires("config").required(false)
         )
         .get_matches()
 }
@@ -420,7 +198,6 @@ fn main() {
             );
         };
 
-        // let mut raw_input;
         let mut command = match values_t!(matches, "s3rs cmd", String) {
             Ok(cmds) => cmds.join(" "),
             Err(_) => String::new(),
